个人中心分成管理入口开关的判断；分成管理页面的调整

+++ Fiu/Classes/NewMain/NewAlliance/Model/Settlement/THNSettlementInfoBalance.h
@@ -28,7 +28,7 @@
 @property (nonatomic, assign) NSInteger productId;
 @property (nonatomic, assign) NSInteger quantity;
 @property (nonatomic, assign) NSInteger skuId;
-@property (nonatomic, assign) NSInteger skuPrice;
+@property (nonatomic, assign) CGFloat skuPrice;
 @property (nonatomic, assign) NSInteger stage;
 @property (nonatomic, strong) NSString * stageLabel;
 @property (nonatomic, assign) NSInteger status;
